+++ tests/test_node.py
@@ -364,7 +364,7 @@ def test_killing_microservice_deletes_rs(cl, client, has_conf):
     has_conf.return_value = False
 
     v1 = MagicMock()
-    client.ExtensionsV1beta1Api.return_value = v1
+    client.AppsV1Api.return_value = v1
 
     body = MagicMock()
     client.V1DeleteOptions.return_value = body